CommonLibrary::IPC::DTO::ProxyTarget

Enum ProxyTarget

pub enum ProxyTarget {

    MainThreadCommands,
    MainThreadConfiguration,
    MainThreadDiagnostics,
    MainThreadDocuments,
    MainThreadExtensionEnablement,
    MainThreadFileSystem,
    MainThreadLanguageFeatures,
    MainThreadLanguages,
    MainThreadOutputService,
    MainThreadSecrets,
    MainThreadStorage,
    MainThreadTerminalService,
    MainThreadWindow,
    MainThreadWebviews,
    MainThreadTelemetry,
    MainThreadWorkspace,
    MainThreadStatusBar,
    MainThreadSourceControlManagement,
    MainThreadTesting,
    MainThreadDebugService,
    MainThreadTaskService,
    MainThreadCustomEditors,
    MainThreadTreeView,
    ExtHostCommands,
    ExtHostConfiguration,
    ExtHostDiagnostics,
    ExtHostDocuments,
    ExtHostExtensionService,
    ExtHostFileSystemInfo,
    ExtHostLanguageFeatures,
    ExtHostLanguages,
    ExtHostOutputService,
    ExtHostStorage,
    ExtHostTerminalService,
    ExtHostEnvironment,
    ExtHostWebviews,
    ExtHostTelemetry,
    ExtHostChatProvider,
    ExtHostExtensionEnablement,
    ExtHostCustomEditors,
    ExtHostQuickInput,
    ExtHostMessageService,
    ExtHostDialogs,
    ExtHostAuthentication,
    ExtHostDebugService,
    ExtHostTaskService,
    ExtHostManagedSockets,
    ExtHostTreeView,
    ExtHostStatusBar,
    ExtHostSourceControlManagement,
    ExtHostTesting,
    ExtHostDebug,

}

An enum that provides strongly-typed identifiers for all services that can be communicated with across the IPC boundary.

This is used to construct fully qualified RPC method names, ensuring that a message sent from one process is routed to the correct service implementation in the other.

Variants

MainThreadCommands

MainThreadConfiguration

MainThreadDiagnostics

MainThreadDocuments

MainThreadExtensionEnablement

MainThreadFileSystem

MainThreadLanguageFeatures

MainThreadLanguages

MainThreadOutputService

MainThreadSecrets

MainThreadStorage

MainThreadTerminalService

MainThreadWindow

MainThreadWebviews

MainThreadTelemetry

MainThreadWorkspace

MainThreadStatusBar

MainThreadSourceControlManagement

MainThreadTesting

MainThreadDebugService

MainThreadTaskService

MainThreadCustomEditors

MainThreadTreeView

ExtHostCommands

ExtHostConfiguration

ExtHostDiagnostics

ExtHostDocuments

ExtHostExtensionService

ExtHostFileSystemInfo

ExtHostLanguageFeatures

ExtHostLanguages

ExtHostOutputService

ExtHostStorage

ExtHostTerminalService

ExtHostEnvironment

ExtHostWebviews

ExtHostTelemetry

ExtHostChatProvider

ExtHostExtensionEnablement

ExtHostCustomEditors

ExtHostQuickInput

ExtHostMessageService

ExtHostDialogs

ExtHostAuthentication

ExtHostDebugService

ExtHostTaskService

ExtHostManagedSockets

ExtHostTreeView

ExtHostStatusBar

ExtHostSourceControlManagement

ExtHostTesting

ExtHostDebug

Implementations

impl ProxyTarget

pub fn GetTargetPrefix(&self) -> String

Returns a string prefix representing the target, used in constructing fully qualified RPC method names (e.g. MainThreadCommands$ExecuteCommand).