CommonLibrary

Module LanguageFeature

LanguageFeature Service

This module defines the abstract contract for all language intelligence services. It includes the main LanguageFeatureProviderRegistry trait, all related Data Transfer Objects (DTOs), and the ActionEffect constructors for every language feature operation. This is the largest and most complex service contract in the application.

Modules

DTO

LanguageFeature DTO Module

LanguageFeatureProviderRegistry

LanguageFeatureProviderRegistry Trait

ProvideCallHierarchy

ProvideCallHierarchy Effect

ProvideCodeActions

ProvideCodeActions Effect

ProvideCodeLenses

ProvideCodeLenses Effect

ProvideCompletions

ProvideCompletions Effect

ProvideDefinition

ProvideDefinition Effect

ProvideDocumentFormatting

ProvideDocumentFormatting Effect

ProvideDocumentHighlights

ProvideDocumentHighlights Effect

ProvideDocumentSymbols

ProvideDocumentSymbols Effect

ProvideFoldingRanges

ProvideFoldingRanges Effect

ProvideHover

ProvideHover Effect

ProvideInlayHints

ProvideInlayHints Effect

ProvideLinkedEditingRanges

ProvideLinkedEditingRanges Effect

ProvideOnTypeFormatting

ProvideOnTypeFormatting Effect

ProvideReferences

ProvideReferences Effect

ProvideRenameEdits

ProvideRenameEdits Effect

ProvideSelectionRanges

ProvideSelectionRanges Effect

ProvideSemanticTokens

ProvideSemanticTokens Effect

ProvideSignatureHelp

ProvideSignatureHelp Effect

ProvideTypeHierarchy

ProvideTypeHierarchy Effect

ProvideWorkspaceSymbols

ProvideWorkspaceSymbols Effect

RegisterProvider

RegisterProvider Effect

UnregisterProvider

UnregisterProvider Effect